feat/ 북,라이브러리북 레퍼지토리 삭제, 리뷰 단일 조회 생성

+++ src/main/kotlin/com/blife/blife/domain/review/service/ReviewService.kt
@@ -2,6 +2,7 @@ package com.blife.blife.domain.review.service
 
 import com.blife.blife.domain.book.repository.BookRepository
 import com.blife.blife.domain.book.repository.LibBookRepository
+import com.blife.blife.domain.member.repository.MemberRepository
 import com.blife.blife.domain.review.dto.BookReviewRequest
 import com.blife.blife.domain.review.dto.BookReviewResponse
 import com.blife.blife.domain.review.dto.LibBookReviewRequest
@@ -20,9 +21,10 @@ class ReviewService(
     private val bookReviewRepository: BookReviewRepository,
     private val libBookReviewRepository: LibBookReviewRepository,
     private val libBookRepository: LibBookRepository,
+    private val memberRepository: MemberRepository,
     private val bookRepository: BookRepository,
 ) {
-    fun getReviewByBookId(bookId: Long, pageable: Pageable): Page<BookReviewResponse> {
+    fun getReviewListByBookId(bookId: Long, pageable: Pageable): Page<BookReviewResponse> {
         val reviewPage = bookReviewRepository.findByBookId(bookId, pageable)
 
 
@@ -35,7 +37,14 @@ class ReviewService(
         }
     }
 
-    fun getReviewByLibBookId(libBookId: Long, pageable: Pageable): Page<LibBookReviewResponse> {
+    fun getReviewByBookId(bookReviewId: Long): BookReviewResponse {
+        val review =
+            bookReviewRepository.findById(bookReviewId).orElseThrow { NoSuchElementException("Book review not found") }
+
+        return review.toResponse()
+    }
+
+    fun getReviewListByLibBookId(libBookId: Long, pageable: Pageable): Page<LibBookReviewResponse> {
         val reviewPage = libBookReviewRepository.findByLibBookId(libBookId, pageable)
 
 
@@ -49,15 +58,22 @@ class ReviewService(
         }
     }
 
+    fun getBookReviewInLib(libBookReviewId: Long): LibBookReviewResponse {
+        val review = libBookReviewRepository.findById(libBookReviewId)
+            .orElseThrow { NoSuchElementException("LibBook review not found") }
+        return review.toLibBookResponse()
+    }
+
+
     @Transactional
-    fun createBookReview(bookId: Long,userId :Long, request: BookReviewRequest): BookReviewResponse {
+    fun createBookReview(bookId: Long, userId: Long, request: BookReviewRequest): BookReviewResponse {
         val book = bookRepository.findById(bookId).orElseThrow { NoSuchElementException("Book not found") }
-        val member =
+        val member = memberRepository.findById(userId).orElseThrow { NoSuchElementException("Member not found") }
         val createdReview = bookReviewRepository.save(
             BookReview(
                 book = book,
                 comment = request.comment,
-                member =
+                member = member,
                 point = request.point,
 
                 )
@@ -66,14 +82,16 @@ class ReviewService(
     }
 
     @Transactional
-    fun createBookReviewInLib(bookId: Long, libraryId: Long, request: LibBookReviewRequest): LibBookReviewResponse {
-        val libBook = libBookRepository.findByBookIdAndLibId(bookId, libraryId)
-            ?: throw NoSuchElementException("LibBook not found")
+    fun createBookReviewInLib(libBookId: Long, userId: Long, request: LibBookReviewRequest): LibBookReviewResponse {
+        val libBook =
+            libBookRepository.findById(libBookId).orElseThrow { NoSuchElementException("Library book not found") }
+        val member = memberRepository.findById(userId).orElseThrow { NoSuchElementException("Member not found") }
+
         val createdReview = libBookReviewRepository.save(
             LibBookReview(
                 libBook = libBook,
                 comment = request.comment,
-                member = TODO(),
+                member = member,
                 point = request.point,
                 status = request.status,
             )
@@ -82,10 +100,14 @@ class ReviewService(
     }
 
     @Transactional
-    fun updateBookReview(bookId: Long, bookReviewId: Long, request: BookReviewRequest): BookReviewResponse {
-        val review = bookReviewRepository.findByBookIdAndId(bookId, bookReviewId)
+    fun updateBookReview(bookReviewId: Long, userId: Long, request: BookReviewRequest): BookReviewResponse {
+        val review =
+            bookReviewRepository.findById(bookReviewId).orElseThrow { NoSuchElementException("Book review not found") }
         if (review == null) {
             throw TODO("no data error")
+        }
+        if (review.member.id != userId) {
+            throw TODO("no authority")
         }
         review.point = request.point
         review.comment = request.comment
@@ -95,12 +117,16 @@ class ReviewService(
 
     @Transactional
     fun updateBookReviewInLib(
-        libBookId: Long, libBookReviewId: Long,
+        libBookReviewId: Long, userId: Long,
         request: LibBookReviewRequest,
     ): LibBookReviewResponse {
-        val review = libBookReviewRepository.findByLibBookIdAndId(libBookId, libBookReviewId)
+        val review = libBookReviewRepository.findById(libBookReviewId)
+            .orElseThrow { NoSuchElementException("LibBook review not found") }
         if (review == null) {
             throw TODO("no data error")
+        }
+        if (review.member.id != userId) {
+            throw TODO("no authority")
         }
         review.point = request.point
         review.comment = request.comment
@@ -111,13 +137,19 @@ class ReviewService(
 
     }
 
-    fun deleteReview(bookReviewId: Long) {
+    fun deleteReview(bookReviewId: Long, userId: Long) {
         val review = bookReviewRepository.findByIdOrNull(bookReviewId) ?: throw TODO("no data error")
+        if (review.member.id != userId) {
+            throw TODO("no authority")
+        }
         return bookReviewRepository.delete(review)
     }
 
-    fun deleteLibBookReview(libBookReviewId: Long) {
+    fun deleteLibBookReview(libBookReviewId: Long, userId: Long) {
         val review = libBookReviewRepository.findByIdOrNull(libBookReviewId) ?: throw TODO("no data error")
+        if (review.member.id != userId) {
+            throw TODO("no authority")
+        }
         return libBookReviewRepository.delete(review)
     }
